Moved headers to appropriate places
