+++ include/freerdp/client/utils/smartcard_cli.h
