+++ include/freerdp/utils/smartcardlogon.h
@@ -42,9 +42,8 @@ typedef struct
 } SmartcardCertInfo;
 
 FREERDP_API BOOL smartcard_enumerateCerts(const rdpSettings* settings, SmartcardCerts** scCert,
-                              DWORD* retCount);
+                                          DWORD* retCount);
 FREERDP_API const SmartcardCertInfo* smartcard_getCertInfo(SmartcardCerts* scCerts, DWORD index);
 FREERDP_API void smartcardCerts_Free(SmartcardCerts* scCert);
-
 
 #endif /* LIBFREERDP_CORE_SMARTCARDLOGON_H */
